+++ starsim/time.py
@@ -322,7 +322,7 @@ class date(pd.Timestamp):
         else:
             if day_round:
                 date = sc.yeartodate(year)
-                timestamp = pd.Timestamp(date).round('d')
+                timestamp = pd.Timestamp(date).round('D')
             else:
                 year_start = pd.Timestamp(year=int(year), month=1, day=1)
                 year_end = pd.Timestamp(year=int(year) + 1, month=1, day=1)
@@ -1801,7 +1801,7 @@ class prob(Rate):
                     errormsg = f'You are trying to convert a unitless probability "{self}" to a different probability with units "{dur}". Are you using ss.prob() instead of e.g. ss.probperyear() or ss.peryear()?'
                     raise TypeError(errormsg)
                 factor = (dur/self.unit)*scale # Main calculation step: cancel units and scale
-                if factor == 1:
+                if sc.isnumber(factor) and factor == 1:
                     return self.value # Avoid expensive calculation and precision issues
                 return 1 - np.exp(-self._rate*factor) # Main use case
         elif sc.isnumber(dur):